+++ libraries/L9110/L9110.h
@@ -3,7 +3,7 @@
 //    FILE: L9110.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2025-09-08
-// VERSION: 0.1.0
+// VERSION: 0.1.1
 // PURPOSE: Arduino library for L9110 motor driver.
 //     URL: https://github.com/RobTillaart/L9110
 //
@@ -13,12 +13,12 @@
 
 #include "Arduino.h"
 
-#define L9110_LIB_VERSION         (F("0.1.0"))
+#define L9110_LIB_VERSION         (F("0.1.1"))
 
 
-#define L9110_STOP                0x00
-#define L9110_REVERSE             0x01
-#define L9110_FORWARD             0x02
+constexpr uint8_t L9110_STOP    = 0x00;
+constexpr uint8_t L9110_REVERSE = 0x01;
+constexpr uint8_t L9110_FORWARD = 0x02;
 
 
 class L9110
@@ -50,12 +50,14 @@ public:
   void forward()
   {
     digitalWrite(_revPin, LOW);
+    if (_delay) delayMicroseconds(_delay);
     digitalWrite(_fwdPin, HIGH);
   }
 
   void reverse()
   {
     digitalWrite(_fwdPin, LOW);
+    if (_delay) delayMicroseconds(_delay);
     digitalWrite(_revPin, HIGH);
   }
 
@@ -73,12 +75,31 @@ public:
     return status;
   }
 
+  void setMicroDelay(uint16_t del = 20)
+  {
+    _delay = del;
+  }
 
-private:
+  uint16_t getMicroDelay()
+  {
+    return _delay;
+  }
+
+
+protected:
   uint8_t _fwdPin;
   uint8_t _revPin;
+  uint16_t _delay = 20;
 };
 
+
+class HG7881 : public L9110
+{
+public:
+  HG7881(uint8_t forwardPin, uint8_t reversePin) : L9110(forwardPin, reversePin)
+  {
+  }
+};
 
 //  -- END OF FILE --