+++ libraries/L9110/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2026-05-13
+- add micro delay for switching forward/reverse.
+- update GitHub actions
+- add derived class HG7881
+- minor edits
+
 ## [0.1.0] - 2023-09-08
 - initial version